fix(suite-native): handle THP correctly for passphrase protection in device settings
